Adding an ReadabilityIndex asbtract class for all indexes
And extends it to all current readability indexes.
